change typedef of exit_t

+++ vm/include/vm.h
@@ -2,10 +2,11 @@
 # define VM_H
 
 # include "libft.h"
+# include "op.h"
 
 # include <fcntl.h>
 # include <stdio.h>
 
-typedef uint8_t exit_t;
+typedef		uint32_t exit_t;
 
 #endif
